api generation script
