+++ scripts/genSolanaAPI.js
@@ -1,19 +1,19 @@
 /* eslint-disable no-console */
 const fs = require('fs');
 const path = require('path');
-const axios = require('axios');
+const fetchEndpoints = require('./utils/apiUtils');
 
 // URL will be changed when api is deployed
-const API_HOST = '';
+const API_HOST = 'solana-gateway.moralis.io';
 const SWAGGER_PATH = '/api-json';
 
 const OUTPUT_DIRECTORY = '../src';
 const OUTPUT_FILENAME = 'MoralisSolanaApi.js';
 
-const BodyParamTypes = {
-  setBody: 'set body',
-  property: 'property',
-};
+// const BodyParamTypes = {
+//   setBody: 'set body',
+//   property: 'property',
+// };
 
 let content = '';
 
@@ -28,7 +28,7 @@ content += `const axios = require('axios');\n`;
 content += `
 class SolanaApi {
   // URL will be changed when api is deployed
-  static baseURL = '';
+  static baseURL = 'https://solana-gateway.moralis.io';
   static BodyParamTypes = {
     setBody: 'set body',
     property: 'property',
@@ -152,93 +152,10 @@ static async apiCall(name, options) {
 
 `;
 
-const fetchSwaggerJson = async () => {
-  const http = await axios.create({ baseURL: `https://${API_HOST}` });
-  const response = await http.get(SWAGGER_PATH);
-  const result = response.data;
-  return result;
-};
-
-const getPathByTag = swaggerJSON => {
-  const pathByTag = {};
-  const pathDetails = {};
-
-  Object.entries(swaggerJSON.paths).map(([pathName, requestData]) => {
-    return Object.entries(requestData).forEach(([method, data]) => {
-      const { tags } = data;
-
-      if (tags.length > 0) {
-        if (!pathByTag[tags[0]]) {
-          pathByTag[tags[0]] = [];
-        }
-        pathByTag[tags[0]].push(data.operationId);
-        pathDetails[data.operationId] = { method, pathName, data };
-      }
-    });
-  });
-
-  return { pathByTag, pathDetails };
-};
-
-const fetchBodyParams = (schema, swaggerJSON, required) => {
-  const body = [];
-  if (schema.$ref) {
-    const splitSchema = schema.$ref.split('/');
-    const schemaKey = splitSchema[splitSchema.length - 1];
-    const component = swaggerJSON.components.schemas[schemaKey];
-    if (component) {
-      Object.entries(component.properties).map(([key, value]) => {
-        return body.push({
-          key,
-          type: component.type === 'array' ? BodyParamTypes.setBody : BodyParamTypes.property,
-          required: component.required.includes(key) ? true : false,
-        });
-      });
-    }
-  } else {
-    body.push({
-      key: 'data',
-      type: BodyParamTypes.setBody,
-      required,
-    });
-  }
-  return body;
-};
-
-const fetchEndpoints = async () => {
-  const swaggerJSON = await fetchSwaggerJson();
-
-  const { pathDetails } = await getPathByTag(swaggerJSON);
-
-  const data = [];
-
-  Object.keys(pathDetails).forEach(x => {
-    const item = pathDetails[x];
-
-    const endpoint = {
-      method: item.method.toUpperCase(),
-      group: item.data.tags[0],
-      name: x,
-      url: item.pathName.split('{').join(':').split('}').join(''),
-      bodyParams: item.data.requestBody
-        ? fetchBodyParams(
-            item.data.requestBody.content['application/json'].schema,
-            swaggerJSON,
-            item.data.requestBody.required
-          )
-        : undefined,
-    };
-
-    data.push(endpoint);
-  });
-
-  return data;
-};
-
 const genSolanaApi = async () => {
   console.log('Start generating automatic SolanaApi code...');
   const wrappers = {};
-  const ENDPOINTS = await fetchEndpoints();
+  const ENDPOINTS = await fetchEndpoints(API_HOST, SWAGGER_PATH);
 
   for (const endpoint of ENDPOINTS) {
     const { group, name, url, method } = endpoint;
