+++ scripts/genWeb3API.js
@@ -1,18 +1,18 @@
 /* eslint-disable no-console */
 const fs = require('fs');
 const path = require('path');
-const axios = require('axios');
+const fetchEndpoints = require('./utils/apiUtils');
 
-const DEEP_INDEX_API_HOST = 'deep-index.moralis.io';
-const DEEP_INDEX_SWAGGER_PATH = '/api-docs/v2/swagger.json';
+const API_HOST = 'deep-index.moralis.io';
+const SWAGGER_PATH = '/api-docs/v2/swagger.json';
 
 const OUTPUT_DIRECTORY = '../src';
 const OUTPUT_FILENAME = 'MoralisWeb3Api.js';
 
-const BodyParamTypes = {
-  setBody: 'set body',
-  property: 'property',
-};
+// const BodyParamTypes = {
+//   setBody: 'set body',
+//   property: 'property',
+// };
 
 let content = '';
 
@@ -146,93 +146,10 @@ static async apiCall(name, options) {
 
 `;
 
-const fetchSwaggerJson = async () => {
-  const http = await axios.create({ baseURL: `https://${DEEP_INDEX_API_HOST}` });
-  const response = await http.get(DEEP_INDEX_SWAGGER_PATH);
-  const result = response.data;
-  return result;
-};
-
-const getPathByTag = swaggerJSON => {
-  const pathByTag = {};
-  const pathDetails = {};
-
-  Object.entries(swaggerJSON.paths).map(([pathName, requestData]) => {
-    return Object.entries(requestData).forEach(([method, data]) => {
-      const { tags } = data;
-
-      if (tags.length > 0) {
-        if (!pathByTag[tags[0]]) {
-          pathByTag[tags[0]] = [];
-        }
-        pathByTag[tags[0]].push(data.operationId);
-        pathDetails[data.operationId] = { method, pathName, data };
-      }
-    });
-  });
-
-  return { pathByTag, pathDetails };
-};
-
-const fetchBodyParams = (schema, swaggerJSON, required) => {
-  const body = [];
-  if (schema.$ref) {
-    const splitSchema = schema.$ref.split('/');
-    const schemaKey = splitSchema[splitSchema.length - 1];
-    const component = swaggerJSON.components.schemas[schemaKey];
-    if (component) {
-      Object.entries(component.properties).map(([key, value]) => {
-        return body.push({
-          key,
-          type: component.type === 'array' ? BodyParamTypes.setBody : BodyParamTypes.property,
-          required: component.required.includes(key) ? true : false,
-        });
-      });
-    }
-  } else {
-    body.push({
-      key: 'abi',
-      type: BodyParamTypes.setBody,
-      required,
-    });
-  }
-  return body;
-};
-
-const fetchEndpoints = async () => {
-  const swaggerJSON = await fetchSwaggerJson();
-
-  const { pathDetails } = await getPathByTag(swaggerJSON);
-
-  const data = [];
-
-  Object.keys(pathDetails).forEach(x => {
-    const item = pathDetails[x];
-
-    const endpoint = {
-      method: item.method.toUpperCase(),
-      group: item.data.tags[0],
-      name: x,
-      url: item.pathName.split('{').join(':').split('}').join(''),
-      bodyParams: item.data.requestBody
-        ? fetchBodyParams(
-            item.data.requestBody.content['application/json'].schema,
-            swaggerJSON,
-            item.data.requestBody.required
-          )
-        : undefined,
-    };
-
-    data.push(endpoint);
-  });
-
-  return data;
-};
-
 const genWebApi = async () => {
   console.log('Start generating automatic Web3API code...');
   const wrappers = {};
-  const ENDPOINTS = await fetchEndpoints();
+  const ENDPOINTS = await fetchEndpoints(API_HOST, SWAGGER_PATH);
 
   for (const endpoint of ENDPOINTS) {
     const { group, name, url, method } = endpoint;
@@ -240,12 +157,6 @@ const genWebApi = async () => {
     wrappers[group][name] = { name, url, method };
   }
 
-  //   console.log(ENDPOINTS);
-  //   console.log(wrappers);
-
-  // ENDPOINTS.forEach(x => {
-  //   wrappers = {};
-  // });
 
   Object.keys(wrappers).forEach(group => {
     content += '\n';
